+++ src/models/title_group_comment.rs
@@ -9,7 +9,7 @@ pub struct TitleGroupComment {
     pub created_at: NaiveDateTime,
     pub updated_at: NaiveDateTime,
     pub created_by_id: i32,
-    pub title_group_id: i32,
+    pub title_group_id: i64,
     pub refers_to_torrent_id: Option<i32>,
     pub answers_to_comment_id: Option<i64>,
 }
@@ -17,7 +17,7 @@ pub struct TitleGroupComment {
 #[derive(Debug, Serialize, Deserialize)]
 pub struct UserCreatedTitleGroupComment {
     pub content: String,
-    pub title_group_id: i32,
+    pub title_group_id: i64,
     pub refers_to_torrent_id: Option<i32>,
     pub answers_to_comment_id: Option<i64>,
 }
